+++ src/main/java/com/kcc/banking/domain/account/mapper/AccountMapper.java
@@ -13,7 +13,7 @@ public interface AccountMapper {
     // 전체 계좌 목록 조회
     List<AccountDetail> findAll();
     // 모달 조회 후 목록 요청
-    List<AccountOfModal> findAccount(SearchAccountOfModal searchAccountOfModal);
+    List<AccountOfModal> findAccountsBySearchOption(SearchAccountOfModal searchAccountOfModal);
     // 계좌 상세정보 조회
     AccountDetail getAccountDetail(String accountId);
